src/main/java/com/simonflarup/gearth/origins/models/outgoing/room/OHCloseTrade.java
package com.simonflarup.gearth.origins.models.outgoing.room;

import com.simonflarup.gearth.origins.models.outgoing.OHServerPacket;
import gearth.protocol.HMessage;
import gearth.protocol.packethandler.shockwave.packets.ShockPacketOutgoing;
import gearth.services.packet_info.PacketInfo;
import gearth.services.packet_info.PacketInfoManager;
import lombok.ToString;

@ToString
public class OHCloseTrade implements OHServerPacket {
    public OHCloseTrade() {
    }

    @Override
    public ShockPacketOutgoing getOutgoingPacket(PacketInfoManager packetInfoManager) {
        PacketInfo packetInfo = packetInfoManager.getPacketInfoFromName(HMessage.Direction.TOSERVER, "TRADE_CLOSE");
        return new ShockPacketOutgoing(packetInfo.getHeaderId());
    }
}
